Move DataHash to picard.util.datahash for general reuse
DataHash (binary data backed by a content-deduplicated temporary file)
was defined in picard.coverart.image but is entirely format-agnostic.
Move it to picard.util.datahash so it can back any large blob that
should not stay resident in memory, not just cover-art image data.

Wording is made neutral (data instead of image); behavior is unchanged.
picard.coverart.image imports it back for CoverArtImage; tagger.py and
the tests now import it from the canonical location.
